pyproject.toml
[project]
name = "waiyacup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django~=4.2",
    "django-browser-reload>=1.18.0",
    "django-debug-toolbar>=5.2.0",
    "django-tailwind~=3.8",
    "django-unfold>=0.59.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "whitenoise>=6.9.0",
]
